@@ -219,11 +218,14 @@
     "raw-loader": "^4.0.2",
     "rimraf": "^3.0.2",
     "spinnies": "^0.5.1",
-    "terser-webpack-plugin": "^5.2.5",
+    "svg-sprite-loader": "^6.0.11",
+    "svg-transform-loader": "^2.0.13",
+    "svgo-loader": "^3.0.0",
+    "terser-webpack-plugin": "^5.3.0",
     "url-loader": "^4.1.1",
-    "vue-loader": "^15.9.8",
+    "vue-loader": "^17.0.0",
     "vue-template-compiler": "^2.6.14",
-    "webpack": "^5.64.2",
+    "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^3.11.2",
     "webpack-hot-middleware": "^2.25.1",
@@ -232,23 +234,25 @@
   "dependencies": {
     "bufferutil": "^4.0.5",
     "crypto-js": "^4.1.1",
-    "electron-log": "^4.4.1",
+    "electron-log": "^4.4.4",
     "electron-store": "^8.0.1",
     "electron-updater": "^4.6.1",
-    "font-list": "git+https://github.com/lyswhut/node-font-list.git#4edbb1933b49a9bac1eedd63a31da16b487fe57d",
-    "http-terminator": "^3.0.3",
+    "font-list": "github:lyswhut/node-font-list#4edbb1933b49a9bac1eedd63a31da16b487fe57d",
+    "http-terminator": "^3.0.4",
     "iconv-lite": "^0.6.3",
     "image-size": "^1.0.0",
     "koa": "^2.13.4",
-    "long": "^5.1.0",
+    "long": "^5.2.0",
+    "mitt": "^3.0.0",
     "needle": "^3.0.0",
     "node-id3": "^0.2.3",
     "request": "^2.88.2",
     "socket.io": "^4.4.0",
+    "sortablejs": "^1.14.0",
     "utf-8-validate": "^5.0.7",
-    "vue": "^2.6.14",
-    "vue-i18n": "^8.26.7",
-    "vue-router": "^3.5.3",
-    "vuex": "^3.6.2"
+    "vue": "^3.2.26",
+    "vue-i18n": "^9.2.0-beta.26",
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2"
   }
 }
