@@ -1,17 +1,15 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 
 import i18n from './plugins/i18n'
 
-import './components'
+import mountComponents from './components'
 
 import App from './App'
 
 import '../common/error'
 
-Vue.config.productionTip = false
 
-new Vue({
-  i18n,
-  el: '#root',
-  render: h => h(App),
-})
+const app = createApp(App)
+app.use(i18n)
+mountComponents(app)
+app.mount('#root')
