@@ -1,17 +1,23 @@
 <template>
-  <Modal :show="visible" @close="handleCancel" @after-leave="afterLeave" :closeBtn="false">
-    <main :class="$style.main">{{message}}</main>
-    <footer :class="$style.footer">
-      <Btn :class="$style.btn" v-if="showCancel" @click="handleCancel">{{cancelBtnText}}</Btn>
-      <Btn :class="$style.btn" @click="handleComfirm">{{confirmBtnText}}</Btn>
-    </footer>
-  </Modal>
+<Modal :show="visible" @close="handleCancel" @after-leave="afterLeave" :closeBtn="false" :teleport="teleport">
+  <main :class="$style.main">{{message}}</main>
+  <footer :class="$style.footer">
+    <Btn :class="$style.btn" v-if="showCancel" @click="handleCancel">{{cancelBtnText}}</Btn>
+    <Btn :class="$style.btn" @click="handleComfirm">{{confirmBtnText}}</Btn>
+  </footer>
+</Modal>
 </template>
 
 <script>
 import Modal from '@renderer/components/material/Modal'
-import Btn from '@renderer/components/material/Btn'
+import Btn from '@renderer/components/base/Btn'
 export default {
+  props: {
+    afterLeave: {
+      type: Function,
+      default: () => {},
+    },
+  },
   components: {
     Modal,
     Btn,
@@ -23,29 +29,25 @@ export default {
       showCancel: false,
       cancelButtonText: '',
       confirmButtonText: '',
+      teleport: '#root',
     }
   },
   computed: {
     cancelBtnText() {
-      return this.cancelButtonText || this.$t('base.cancel_button_text')
+      return this.cancelButtonText || this.$t('cancel_button_text')
     },
     confirmBtnText() {
-      return this.confirmButtonText || this.$t('base.confirm_button_text')
+      return this.confirmButtonText || this.$t('confirm_button_text')
     },
   },
-  beforeDestroy() {
+  beforeUnmount() {
     const el = this.$el
     el.parentNode.removeChild(el)
   },
   methods: {
-    afterLeave(el, done) {
-      this.$destroy()
-    },
     handleCancel() {
-
     },
     handleComfirm() {
-
     },
   },
 }
