@@ -9,7 +9,7 @@ const baseConfig = require('./webpack.config.base')
 
 const { dependencies } = require('../../package.json')
 
-let whiteListedModules = ['vue']
+let whiteListedModules = ['vue', 'vue-router', 'vuex', 'vue-i18n']
 
 
 module.exports = merge(baseConfig, {
@@ -31,9 +31,12 @@ module.exports = merge(baseConfig, {
       'process.env': {
         NODE_ENV: '"production"',
       },
+      __VUE_OPTIONS_API__: 'true',
+      __VUE_PROD_DEVTOOLS__: 'false',
     }),
   ],
   optimization: {
+    minimize: false,
     minimizer: [
       new TerserPlugin(),
       new CssMinimizerPlugin(),
