@@ -6,7 +6,7 @@
  */
 const electron = require('electron')
 const electronDebug = require('electron-debug')
-const { default: installExtension, VUEJS_DEVTOOLS } = require('electron-devtools-installer')
+const { default: installExtension, VUEJS3_DEVTOOLS } = require('electron-devtools-installer')
 // Install `electron-debug` with `devtron`
 electronDebug({
   showDevTools: true,
@@ -15,7 +15,7 @@ electronDebug({
 
 // Install `vue-devtools`
 electron.app.on('ready', () => {
-  installExtension(VUEJS_DEVTOOLS)
+  installExtension(VUEJS3_DEVTOOLS)
     .then(name => console.log(`Added Extension:  ${name}`))
     .catch(err => console.log('An error occurred: ', err))
 })
