@@ -1,10 +1,11 @@
-import Vue from 'vue'
+import mitt from 'mitt'
 import keyBind from '../utils/keyBind'
 import { rendererOn, rendererSend, NAMES, rendererInvoke } from '../../common/ipc'
 import { base as baseName, sync as syncName } from './names'
 import { common as hotKeyNamesCommon } from '../../common/hotKey'
+import { sync } from '@renderer/core/share'
 
-const eventHub = window.eventHub = new Vue()
+const eventHub = window.eventHub = mitt()
 
 window.isEditingHotKey = false
 let appHotKeyConfig = {
@@ -18,29 +19,29 @@ rendererInvoke(NAMES.mainWindow.get_hot_key).then(({ local, global }) => {
   }
 })
 
-eventHub.$on(baseName.bindKey, () => {
+eventHub.on(baseName.bindKey, () => {
   keyBind.bindKey((key, type, event, keys) => {
     // console.log(`key_${key}_${type}`)
-    eventHub.$emit(baseName.key_down, { event, keys, key, type })
+    eventHub.emit(baseName.key_down, { event, keys, key, type })
     // console.log(event, key)
     if (!window.isEditingHotKey && appHotKeyConfig.local.enable && appHotKeyConfig.local.keys[key]) {
       if (type == 'up') return
 
       // 软件内快捷键的最小化触发时
       // 如果已启用托盘，则隐藏程序，否则最小化程序 https://github.com/lyswhut/lx-music-desktop/issues/603
-      if (appHotKeyConfig.local.keys[key].action == hotKeyNamesCommon.min.action && global.appSetting.tray.isToTray) {
-        eventHub.$emit(hotKeyNamesCommon.hide_toggle.action)
+      if (appHotKeyConfig.local.keys[key].action == hotKeyNamesCommon.min.action && global.appSetting.tray.isShow) {
+        eventHub.emit(hotKeyNamesCommon.hide_toggle.action)
         return
       }
 
-      eventHub.$emit(appHotKeyConfig.local.keys[key].action)
+      eventHub.emit(appHotKeyConfig.local.keys[key].action)
       return
     }
-    eventHub.$emit(`key_${key}_${type}`, { event, keys, key, type })
+    eventHub.emit(`key_${key}_${type}`, { event, keys, key, type })
   })
   registerCommonEvents()
 })
-eventHub.$on(baseName.unbindKey, () => {
+eventHub.on(baseName.unbindKey, () => {
   keyBind.unbindKey()
   unregisterCommonEvents()
 })
@@ -49,31 +50,35 @@ const registerMin = () => rendererSend(NAMES.mainWindow.min)
 const registerMinToggle = () => rendererSend(NAMES.mainWindow.min_toggle)
 const registerHideToggle = () => rendererSend(NAMES.mainWindow.hide_toggle)
 
-eventHub.$on(baseName.min, registerMin)
-eventHub.$on(baseName.max, () => rendererSend(NAMES.mainWindow.max))
-eventHub.$on(baseName.close, () => rendererSend(NAMES.mainWindow.close))
+const setClearDownKeys = () => keyBind.clearDownKeys()
+
+eventHub.on(baseName.min, registerMin)
+eventHub.on(baseName.max, () => rendererSend(NAMES.mainWindow.max))
+eventHub.on(baseName.close, () => rendererSend(NAMES.mainWindow.close))
+
+eventHub.on(baseName.setClearDownKeys, setClearDownKeys)
 
 const registerCommonEvents = () => {
-  eventHub.$on(hotKeyNamesCommon.close.action, registerQuit)
-  eventHub.$on(hotKeyNamesCommon.min.action, registerMin)
-  eventHub.$on(hotKeyNamesCommon.min_toggle.action, registerMinToggle)
-  eventHub.$on(hotKeyNamesCommon.hide_toggle.action, registerHideToggle)
+  eventHub.on(hotKeyNamesCommon.close.action, registerQuit)
+  eventHub.on(hotKeyNamesCommon.min.action, registerMin)
+  eventHub.on(hotKeyNamesCommon.min_toggle.action, registerMinToggle)
+  eventHub.on(hotKeyNamesCommon.hide_toggle.action, registerHideToggle)
 }
 const unregisterCommonEvents = () => {
-  eventHub.$off(hotKeyNamesCommon.close.action, registerQuit)
-  eventHub.$off(hotKeyNamesCommon.min.action, registerMin)
-  eventHub.$off(hotKeyNamesCommon.min_toggle.action, registerMinToggle)
-  eventHub.$off(hotKeyNamesCommon.hide_toggle.action, registerHideToggle)
+  eventHub.off(hotKeyNamesCommon.close.action, registerQuit)
+  eventHub.off(hotKeyNamesCommon.min.action, registerMin)
+  eventHub.off(hotKeyNamesCommon.min_toggle.action, registerMinToggle)
+  eventHub.off(hotKeyNamesCommon.hide_toggle.action, registerHideToggle)
 }
 
 rendererOn(NAMES.mainWindow.focus, () => {
   keyBind.clearDownKeys()
-  eventHub.$emit(baseName.focus)
+  eventHub.emit(baseName.focus)
 })
 rendererOn(NAMES.mainWindow.key_down, (event, { type, key }) => {
   // console.log(appHotKeyConfig.global.keys[key])
   if (appHotKeyConfig.global.keys[key]) {
-    window.eventHub.$emit(appHotKeyConfig.global.keys[key].action)
+    window.eventHub.emit(appHotKeyConfig.global.keys[key].action)
   }
 })
 rendererOn(NAMES.mainWindow.set_hot_key_config, (event, config) => {
@@ -83,20 +88,20 @@ rendererOn(NAMES.mainWindow.set_hot_key_config, (event, config) => {
   for (const type of Object.keys(config)) {
     window.appHotKeyConfig[type] = config[type]
   }
-  window.eventHub.$emit(baseName.set_hot_key_config, config)
+  window.eventHub.emit(baseName.set_hot_key_config, config)
 })
 
 rendererOn(NAMES.mainWindow.sync_action_list, (event, { action, data }) => {
-  window.eventHub.$emit(syncName.handle_action_list, { action, data })
+  window.eventHub.emit(syncName.handle_action_list, { action, data })
 })
-eventHub.$on(syncName.send_action_list, ({ action, data }) => {
-  if (!window.globalObj.sync.enable) return
+eventHub.on(syncName.send_action_list, ({ action, data }) => {
+  if (!sync.enable) return
   rendererSend(NAMES.mainWindow.sync_action_list, { action, data })
 })
 rendererOn(NAMES.mainWindow.sync_list, (event, { action, data }) => {
-  window.eventHub.$emit(syncName.handle_sync_list, { action, data })
+  window.eventHub.emit(syncName.handle_sync_list, { action, data })
 })
-eventHub.$on(syncName.send_sync_list, ({ action, data }) => {
-  if (!window.globalObj.sync.enable) return
+eventHub.on(syncName.send_sync_list, ({ action, data }) => {
+  if (!sync.enable) return
   rendererSend(NAMES.mainWindow.sync_list, { action, data })
 })
