@@ -1,10 +1,12 @@
 import tips from './Tips'
 import { debounce } from '../../utils'
+import { base as eventBaseName } from '@renderer/event/names'
 
 let instance
 let prevTips
 let prevX = 0
 let prevY = 0
+let isDraging = false
 
 const getTips = el =>
   el
@@ -16,6 +18,7 @@ const getTips = el =>
     : null
 
 const showTips = debounce(event => {
+  if (isDraging) return
   let msg = getTips(event.target)
   if (!msg) return
   prevTips = msg
@@ -26,11 +29,12 @@ const showTips = debounce(event => {
       top: event.y + 12,
       left: event.x + 8,
     },
-  })
-  instance.$on('beforeClose', closeInstance => {
-    if (instance !== closeInstance) return
-    prevTips = null
-    instance = null
+  }, {
+    beforeClose(closeInstance) {
+      if (instance !== closeInstance) return
+      prevTips = null
+      instance = null
+    },
   })
 }, 400)
 
@@ -45,6 +49,7 @@ const setTips = tips => {
 }
 
 const updateTips = event => {
+  if (isDraging) return
   if (!instance) return showTips(event)
   setTimeout(() => {
     let msg = getTips(event.target)
@@ -55,7 +60,7 @@ const updateTips = event => {
 }
 
 document.body.addEventListener('mousemove', event => {
-  if (event.x == prevX && event.y == prevY) return
+  if ((event.x == prevX && event.y == prevY) || isDraging) return
   prevX = event.x
   prevY = event.y
   hideTips()
@@ -65,3 +70,14 @@ document.body.addEventListener('mousemove', event => {
 document.body.addEventListener('click', updateTips)
 
 document.body.addEventListener('contextmenu', updateTips)
+
+window.eventHub.on(eventBaseName.focus, () => {
+  hideTips()
+})
+window.eventHub.on(eventBaseName.dragStart, () => {
+  isDraging = true
+  hideTips()
+})
+window.eventHub.on(eventBaseName.dragEnd, () => {
+  isDraging = false
+})
