@@ -1,3 +1,10 @@
 // import './axios'
-import './Dialog'
+import dialog from './Dialog'
 import './Tips'
+import svgIcon from './SvgIcon'
+
+export default app => {
+  app.use(dialog)
+
+  svgIcon(app)
+}
