@@ -1,14 +1,15 @@
 import music from '../utils/music'
+import { themes, windowSizeList } from '@renderer/core/share'
 
 export default {
   theme(state) {
-    let theme = state.themes.find(theme => theme.id == state.setting.themeId)
-    return (theme && theme.class) || ''
+    let theme = themes.find(theme => theme.id == state.setting.themeId)
+    return (theme && theme.className) || ''
   },
   themes(state) {
     return {
       active: state.setting.themeId,
-      list: state.themes,
+      list: themes,
     }
   },
   source(state) {
@@ -22,7 +23,7 @@ export default {
     }
   },
   sourceNames(state) {
-    let prefix = 'store.state.source_'
+    let prefix = 'source_'
     if (state.setting.sourceNameType == 'alias') prefix += 'alias_'
     const sources = {}
     for (const source of music.sources) {
@@ -46,10 +47,13 @@ export default {
   route(state) {
     return state.route
   },
-  windowSizeList(state) {
-    return state.windowSizeList
+  windowSizeList() {
+    return windowSizeList
   },
   windowSizeActive(state) {
-    return state.windowSizeList.find(i => i.id === state.setting.windowSizeId) || state.windowSizeList[0]
+    return windowSizeList.find(i => i.id === state.setting.windowSizeId) || windowSizeList[0]
+  },
+  pactModalVisible(state) {
+    return !state.setting.isAgreePact
   },
 }
