@@ -1,5 +1,4 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { createStore } from 'vuex'
 // import { createPersistedState, createSharedMutations } from 'vuex-electron'
 
 import defaultState from './state'
@@ -8,11 +7,9 @@ import modules from './modules'
 import getters from './getters'
 import actions from './actions'
 
-Vue.use(Vuex)
-
 const isDev = process.env.NODE_ENV === 'development'
 
-const store = new Vuex.Store({
+const store = createStore({
   strict: isDev,
   state: defaultState,
   modules,
