@@ -1,4 +1,5 @@
 import music from '../../utils/music'
+import { markRawList } from '@renderer/utils/vueTools'
 
 const sources = []
 const sourceList = {}
@@ -150,7 +151,7 @@ const mutations = {
   },
   setList(state, datas) {
     let source = state.sourceList[datas.source]
-    source.list = datas.list
+    source.list = markRawList(datas.list)
     source.total = datas.total
     source.allPage = datas.allPage
     source.page = datas.page
@@ -164,7 +165,7 @@ const mutations = {
     for (const source of results) {
       state.sourceMaxPage[source.source] = source.allPage
       if (source.allPage < page) continue
-      list.push(...source.list)
+      list.push(...markRawList(source.list))
       pages.push(source.allPage)
       total += source.total
       // limit = Math.max(source.limit, limit)
