@@ -24,21 +24,6 @@ export default {
     if (sortId != null) state.setting.songList.sortId = sortId
     if (source != null) state.setting.songList.source = source
   },
-  setNewVersion(state, val) {
-    state.version.newVersion = val
-  },
-  setDownloadProgress(state, info) {
-    state.version.downloadProgress = info
-  },
-  setVersionModalVisible(state, { isShow, isError, isDownloaded, isTimeOut, isDownloading, isUnknow, isLatestVer }) {
-    if (isError !== undefined) state.version.isError = isError
-    if (isTimeOut !== undefined) state.version.isTimeOut = isTimeOut
-    if (isDownloading !== undefined) state.version.isDownloading = isDownloading
-    if (isDownloaded !== undefined) state.version.isDownloaded = isDownloaded
-    if (isUnknow !== undefined) state.version.isUnknow = isUnknow
-    if (isLatestVer !== undefined) state.version.isLatestVer = isLatestVer
-    if (isShow !== undefined) state.version.showModal = isShow
-  },
   setIgnoreVersion(state, version) {
     state.setting.ignoreVersion = version
   },
@@ -63,5 +48,14 @@ export default {
   },
   setDesktopLyricConfig(state, config) {
     state.setting.desktopLyric = Object.assign(state.setting.desktopLyric, config)
+  },
+  setApiSource(state, apiSource) {
+    state.setting.apiSource = apiSource
+  },
+  setProxyEnable(state, val) {
+    state.setting.network.proxy.enable = val
+  },
+  setAudioVisualization(state, val) {
+    state.setting.player.audioVisualization = val
   },
 }
